pyproject.toml
[build-system]
requires = ["setuptools >= 64", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "ClusterClue"
version = "0.1.0"
license = "AGPL-3.0-or-later"
description = "Create and detect biosynthetic subcluster motifs in BGCs"
readme = "README.md"
requires-python = ">=3.12,<3.13"
authors = [
    { name = "Annette Lien", email = "annette.lien@posteo.de" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
    "biopython",
    "pyhmmer",
    "matplotlib",
    "networkx",
    "numpy",
    "gensim==4.3.3",
    "pyLDAvis",
    "pandas",
    "seaborn",
    "statsmodels",
    "sympy",
    "scikit-learn",
    "faiss-cpu",
    "cdlib",
    "python-igraph",
    "leidenalg",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov"
]
jupyter = [
    "jupyter",
    "notebook",
    "rdkit",
]

[project.scripts]
clusterclue = "clusterclue.cli:main"

[tool.setuptools.dynamic]
version = {attr = "clusterclue.__version__"}

[tool.setuptools.packages.find]
# find all packages
where = ["."]
include = ["*"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"clusterclue" = ["data/*"]